Initial changes to remove user with notification.

+++ src/maidsafe/lifestuff/session.cc
@@ -86,7 +86,9 @@ PublicIdContactMap Session::GetAllContacts(ContactStatus status) {
   auto it(contact_handler_map_.begin());
   for (; it != contact_handler_map_.end(); ++it) {
     result[(*it).first] = std::set<std::string>();
-    (*it).second->OrderedContacts(&contacts, kAlphabetical, status);
+    (*it).second->OrderedContacts(&contacts,
+                                  kAlphabetical,
+                                  static_cast<uint16_t>(status));
     for (auto item(contacts.begin()); item != contacts.end(); ++item)
       result[(*it).first].insert((*item).public_username);
   }
